Refactor Periscope GUI into separate modules
